+++ sgtl/graph.py
@@ -1,13 +1,13 @@
 """
 Provides the Graph object, which is our core representation of a graph within the SGTL library.
 """
-import scipy as sp
+import math
+import scipy
 import scipy.sparse
 import numpy as np
-import math
 
 
-class Graph(object):
+class Graph:
     """
     Represents a graph. We keep things very simple - a graph is represented by its sparse adjacency matrix.
 
@@ -30,33 +30,52 @@ class Graph(object):
         :param adj_mat: A sparse scipy matrix.
         """
         # The graph is represented by the sparse adjacency matrix. We store the adjacency matrix in two sparse formats.
+        # We can assume that there are no non-zero entries in the stored adjacency matrix.
         self.adj_mat = adj_mat.tocsr()
+        self.adj_mat.eliminate_zeros()
         self.lil_adj_mat = adj_mat.tolil()
 
-        # For convenience, and to speed up operations on the graph, we precompute various pieces of information about
-        # the graph.
-
-        # Store the degrees of the vertices in the graph.
+        # For convenience, and to speed up operations on the graph, we precompute the degrees of the vertices in the
+        # graph.
         self.degrees = adj_mat.sum(axis=0).tolist()[0]
         self.inv_degrees = list(map(lambda x: 1 / x if x != 0 else 0, self.degrees))
         self.sqrt_degrees = list(map(math.sqrt, self.degrees))
         self.inv_sqrt_degrees = list(map(lambda x: 1 / x if x > 0 else 0, self.sqrt_degrees))
 
-        # Store the number of edges and vertices in the graph.
-        self.num_vertices = self.adj_mat.shape[0]
-        self.num_edges = round(sum(self.degrees) / 2)
+    def number_of_vertices(self) -> int:
+        """The number of vertices in the graph."""
+        return self.adj_mat.shape[0]
+
+    def total_volume(self) -> float:
+        """The total volume of the graph."""
+        return (sum(self.degrees) + self._volume_of_self_loops()) / 2
+
+    def _number_of_self_loops(self) -> int:
+        """Get the number of self-loops in the graph."""
+        return np.count_nonzero(self.adj_mat.diagonal())
+
+    def _volume_of_self_loops(self) -> float:
+        """Get the total weight of all self-loops in the graph."""
+        return float(np.sum(self.adj_mat.diagonal()))
+
+    def number_of_edges(self) -> int:
+        """The number of edges in the graph, ignoring any weights."""
+        return int((self.adj_mat.nnz + self._number_of_self_loops()) / 2)
 
     def degree_matrix(self):
         """Construct the diagonal degree matrix of the graph."""
-        return sp.sparse.spdiags(self.degrees, [0], self.num_vertices, self.num_vertices, format="csr")
+        return scipy.sparse.spdiags(
+            self.degrees, [0], self.number_of_vertices(), self.number_of_vertices(), format="csr")
 
     def inverse_degree_matrix(self):
         """Construct the inverse of the diagonal degree matrix of the graph."""
-        return sp.sparse.spdiags(self.inv_degrees, [0], self.num_vertices, self.num_vertices, format="csr")
+        return scipy.sparse.spdiags(self.inv_degrees, [0], self.number_of_vertices(), self.number_of_vertices(),
+                                    format="csr")
 
     def inverse_sqrt_degree_matrix(self):
         """Construct the square root of the inverse of the diagonal degree matrix of the graph."""
-        return sp.sparse.spdiags(self.inv_sqrt_degrees, [0], self.num_vertices, self.num_vertices, format="csr")
+        return scipy.sparse.spdiags(self.inv_sqrt_degrees, [0], self.number_of_vertices(), self.number_of_vertices(),
+                                    format="csr")
 
     def laplacian_matrix(self):
         """
@@ -159,11 +178,11 @@ class Graph(object):
         return 1 - 2 * self.weight(vertex_set_l, vertex_set_r) / self.volume(vertex_set_l + vertex_set_r)
 
 
-def complete_graph(n: int) -> Graph:
+def complete_graph(number_of_vertices: int) -> Graph:
     """
     Construct the complete unweighted graph on :math:`n` vertices.
 
-    :param n: The number of vertices in the graph.
+    :param number_of_vertices: The number of vertices in the graph.
     :return: The complete graph on :math:`n` vertices, as a `Graph` object.
     :raises ValueError: if the number of vertices is not a positive integer.
 
@@ -178,19 +197,19 @@ def complete_graph(n: int) -> Graph:
               [1., 1., 1., 0.]])
 
     """
-    if n <= 0:
+    if number_of_vertices <= 0:
         raise ValueError("The graph must contain at least one vertex.")
 
     # Generate the complete adjacency matrix - we generate a dense matrix first
-    adj_mat = sp.sparse.csr_matrix(np.ones((n, n)) - np.eye(n))
+    adj_mat = scipy.sparse.csr_matrix(np.ones((number_of_vertices, number_of_vertices)) - np.eye(number_of_vertices))
     return Graph(adj_mat)
 
 
-def cycle_graph(n: int) -> Graph:
+def cycle_graph(number_of_vertices: int) -> Graph:
     """
     Construct the unweighted cycle graph on :math:`n` vertices.
 
-    :param n: The number of vertices in the graph
+    :param number_of_vertices: The number of vertices in the graph
     :return: The cycle graph on :math:`n` vertices, as a `Graph` object.
     :raises ValueError: if the number of vertices is not a positive integer.
 
@@ -206,19 +225,23 @@ def cycle_graph(n: int) -> Graph:
               [1., 0., 0., 1., 0.]])
 
     """
-    if n <= 0:
+    if number_of_vertices <= 0:
         raise ValueError("The graph must contain at least one vertex.")
 
     # Generate the cycle graph adjacency matrix
-    adj_mat = sp.sparse.diags([np.ones(n-1), np.ones(n-1), np.ones(1), np.ones(1)], [-1, 1, (n-1), (1-n)])
+    adj_mat = scipy.sparse.diags([np.ones(number_of_vertices - 1),
+                                  np.ones(number_of_vertices - 1),
+                                  np.ones(1),
+                                  np.ones(1)],
+                                 [-1, 1, (number_of_vertices - 1), (1 - number_of_vertices)])
     return Graph(adj_mat)
 
 
-def star_graph(n: int) -> Graph:
+def star_graph(number_of_vertices: int) -> Graph:
     """
     Construct the unweighted star graph on :math:`n` vertices.
 
-    :param n: The number of vertices in the graph
+    :param number_of_vertices: The number of vertices in the graph
     :return: The star graph on :math:`n` vertices, as a `Graph` object.
     :raises ValueError: if the number of vertices is not a positive integer.
 
@@ -233,22 +256,22 @@ def star_graph(n: int) -> Graph:
               [1., 0., 0., 0.]])
 
     """
-    if n <= 0:
+    if number_of_vertices <= 0:
         raise ValueError("The graph must contain at least one vertex.")
 
     # Generate the cycle graph adjacency matrix
-    adj_mat = sp.sparse.lil_matrix((n, n))
-    for i in range(1, n):
+    adj_mat = scipy.sparse.lil_matrix((number_of_vertices, number_of_vertices))
+    for i in range(1, number_of_vertices):
         adj_mat[0, i] = 1
         adj_mat[i, 0] = 1
     return Graph(adj_mat)
 
 
-def path_graph(n: int) -> Graph:
+def path_graph(number_of_vertices: int) -> Graph:
     """
     Construct the unweighted path graph on :math:`n` vertices.
 
-    :param n: The number of vertices in the graph
+    :param number_of_vertices: The number of vertices in the graph
     :return: The path graph on :math:`n` vertices, as a `Graph` object.
     :raises ValueError: if the number of vertices is not a positive integer.
 
@@ -264,9 +287,9 @@ def path_graph(n: int) -> Graph:
               [0., 0., 0., 1., 0.]])
 
     """
-    if n <= 0:
+    if number_of_vertices <= 0:
         raise ValueError("The graph must contain at least one vertex.")
 
     # Generate the cycle graph adjacency matrix
-    adj_mat = sp.sparse.diags([np.ones(n-1), np.ones(n-1)], [-1, 1])
+    adj_mat = scipy.sparse.diags([np.ones(number_of_vertices - 1), np.ones(number_of_vertices - 1)], [-1, 1])
     return Graph(adj_mat)