+++ tests/test_graph.py
@@ -34,8 +34,8 @@ def test_graph_constructor():
     graph = sgtl.Graph(C4_ADJ_MAT)
 
     # The graph has 4 vertices and 4 edges
-    assert graph.num_vertices == 4
-    assert graph.num_edges == 4
+    assert graph.number_of_vertices() == 4
+    assert graph.number_of_edges() == 4
 
     # Check the vertex degrees
     for i in range(4):
@@ -46,8 +46,8 @@ def test_graph_constructor():
 
     # Now, try creating the complete graph on 6 vertices.
     graph = sgtl.Graph(K6_ADJ_MAT)
-    assert graph.num_vertices == 6
-    assert graph.num_edges == 15
+    assert graph.number_of_vertices() == 6
+    assert graph.number_of_edges() == 15
     for i in range(4):
         assert graph.degrees[i] == 5
         assert graph.inv_degrees[i] == 1/5
@@ -56,8 +56,8 @@ def test_graph_constructor():
 
     # Now, try the barbell graph
     graph = sgtl.Graph(BARBELL5_ADJ_MAT)
-    assert graph.num_vertices == 10
-    assert graph.num_edges == 21
+    assert graph.number_of_vertices() == 10
+    assert graph.number_of_edges() == 21
     assert graph.degrees[2] == 4
     assert graph.degrees[4] == 5
 
@@ -68,7 +68,7 @@ def test_complete_graph():
     graph = sgtl.graph.complete_graph(n)
     expected_adjacency_matrix = sp.sparse.csr_matrix([[0, 1, 1, 1], [1, 0, 1, 1], [1, 1, 0, 1], [1, 1, 1, 0]])
 
-    assert graph.num_vertices == 4
+    assert graph.number_of_vertices() == 4
     adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
@@ -90,7 +90,7 @@ def test_cycle_graph():
                                                       [0, 0, 1, 0, 1],
                                                       [1, 0, 0, 1, 0]])
 
-    assert graph.num_vertices == 5
+    assert graph.number_of_vertices() == 5
     adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
@@ -112,7 +112,7 @@ def test_star_graph():
                                                       [1, 0, 0, 0, 0],
                                                       [1, 0, 0, 0, 0]])
 
-    assert graph.num_vertices == 5
+    assert graph.number_of_vertices() == 5
     adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
@@ -134,7 +134,7 @@ def test_path_graph():
                                                       [0, 0, 1, 0, 1],
                                                       [0, 0, 0, 1, 0]])
 
-    assert graph.num_vertices == 5
+    assert graph.number_of_vertices() == 5
     adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
@@ -190,3 +190,50 @@ def test_symmetry():
     lap_mat = big_graph.laplacian_matrix()
     lap_mat_dense = lap_mat.toarray()
     assert np.allclose(lap_mat_dense, lap_mat_dense.T)
+
+
+def test_num_edges():
+    # Generate a known graph
+    graph = sgtl.Graph(BARBELL5_ADJ_MAT)
+
+    # Check the number of edges in the graph
+    assert graph.number_of_vertices() == 10
+    assert graph.number_of_edges() == 21
+    assert graph.total_volume() == 21
+
+    # Now create a weighted graph and check the number of edges method.
+    adjacency_matrix = scipy.sparse.csr_matrix([[0, 2, 0, 1],
+                                                [2, 0, 3, 0],
+                                                [0, 3, 0, 1],
+                                                [1, 0, 1, 0]])
+    graph = sgtl.Graph(adjacency_matrix)
+    assert graph.number_of_vertices() == 4
+    assert graph.number_of_edges() == 4
+    assert graph.total_volume() == 7
+
+    # Test the number of edges for a graph with self-loops
+    adjacency_matrix = scipy.sparse.csr_matrix([[0, 2, 0, 1],
+                                                [2, 2, 3, 0],
+                                                [0, 3, 0, 1],
+                                                [1, 0, 1, 0]])
+    graph = sgtl.Graph(adjacency_matrix)
+    assert graph.number_of_vertices() == 4
+    assert graph.number_of_edges() == 5
+    assert graph.total_volume() == 9
+
+    # Add more self-loops to a graph
+    adjacency_matrix = scipy.sparse.csr_matrix([[0, 1, 1, 1, 1, 0, 0, 0, 0, 0],
+                                                [1, 3, 1, 1, 1.5, 0, 0, 0, 0, 0],
+                                                [1, 1, 0, 1, 1, 0, 0, 0, 0, 0],
+                                                [1, 1.5, 1, 0, 1, 0, 0, 0, 0, 0],
+                                                [1, 1, 1, 1, 0.5, 1, 0, 0, 0, 0],
+                                                [0, 0, 0, 0, 1, 0, 1, 1, 1, 1],
+                                                [0, 0, 0, 0, 0, 1, 0.5, 1, 2.5, 1],
+                                                [0, 0, 0, 0, 0, 1, 1, 0, 1, 1],
+                                                [0, 0, 0, 0, 0, 1, 2.5, 1, 0, 0.5],
+                                                [0, 0, 0, 0, 0, 1, 1, 1, 0.5, 0],
+                                                ])
+    graph = sgtl.Graph(adjacency_matrix)
+    assert graph.number_of_vertices() == 10
+    assert graph.number_of_edges() == 24
+    assert graph.total_volume() == 26.5